Always honor syslog environment variables
The library allows you to configure logging to syslog with
SPOTIFY_SYSLOG_HOST and SPOTIFY_SYSLOG_PORT, but these would only
be honored if --syslog, --syslogHost or --syslogPort was passed in
to the JewelCliLoggingConfigurator. This commit makes it so that the
environment variables will be used even if the command line args are
not specified. This makes it easier to use within docker containers,
so users can simply pass the environment variables to the container,
without worrying about the other arguments.

The configureSyslogDefaults static method has also been modified to
honor the environment variables if present, instead of just defaulting
to localhost.