Make configureDefaults honor syslog environment variables
Before SPOTIFY_SYSLOG_HOST and SPOTIFY_SYSLOG_PORT only had an effect if
configureSyslogDefaults was called. That meant callers would need to check
for the environment variables themselves, and then call the approriate
method. This defeats the point of the env vars, which was to have the
library configure itself. Now, callers can just call configureDefaults,
and if either env var is set, configureSyslogDefaults will be called
automatically.